+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+- Updated System.Text.Json dependency (CVE-2024-43485)
+
 ### 5.3.5
 - [Base] added IsEmpty/IsEmptyOrNull overloads for Array/ICollection with efficient implementation
 - [Base] added FindValue and FindNeighboursV to SortedSet