+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 - Added Ray3.Transformed overloads
 - Added Ray3.Normalized
 - Added Vec.AngleBetweenSigned for 2D vectors
+- Added Fun.WrapToPi
 
 ### 5.3.5
 - [Base] added IsEmpty/IsEmptyOrNull overloads for Array/ICollection with efficient implementation